In-class work and refactored the project.
