+++ checkers.html
@@ -5,6 +5,10 @@
   </head>
   <body>
     <!-- Game here -->
+    <script src="./Utility.js" type="text/javascript"></script>
+    <script src="./GameDataLogic.js" type="text/javascript"></script>
+    <script src="./CheckerBoards/CanvasBoard.js" type="text/javascript"></script>
+    <script src="./CheckerBoards/HTMLBoard.js" type="text/javascript"></script>
     <script src="./checkers.js" type="text/javascript"></script>
   </body>
 </html>
